не работает addMorningPlaylist(id)
